feat: :sparkles: Update AlumnoCreditosReport type and refactor imports across components
