+++ src/types/credito.ts
@@ -27,13 +27,3 @@ export interface UpdateCreditoDto {
     act_id?: number;
     cred_fecha?: string;
 }
-
-export type AlumnoCreditosReport = {
-    alumno: Alumno;
-    totalCreditos: number;
-    creditos: {
-        docente: string;
-        actividad: string;
-        fecha: string;
-    }[];
-}
